cnn14/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
birdnet==0.1.7
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.45.1
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159252575002/conda-bld/mkl_fft_1761592901585/work
mkl_random @ file:///home/task_176159293813909/conda-bld/mkl_random_1761592952512/work
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.3
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opt_einsum==3.4.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==4.25.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
PyWavelets @ file:///croot/pywavelets_1737133681996/work
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
timm==1.0.20
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchlibrosa==0.1.0
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
yarl==1.22.0
